+++ src/DonorDashboards/resources/js/app/components/annual-receipt-table/annual-receipt-row/index.js
@@ -1,5 +1,6 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-const { __ } = wp.i18n;
+import { __ } from '@wordpress/i18n';
+;
 
 const AnnualReceiptRow = ( { annualReceipt } ) => {
 	const { year, amount, count, statementUrl } = annualReceipt[ 1 ];