+++ src/DonorDashboards/resources/js/app/components/annual-receipt-table/index.js
@@ -1,6 +1,7 @@
 import { Fragment, useState } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-const { __ } = wp.i18n;
+import { __ } from '@wordpress/i18n';
+;
 
 import Table from '../table';
 import AnnualReceiptRow from './annual-receipt-row';