+++ src/DonorDashboards/resources/js/app/components/subscription-manager/index.js
@@ -3,7 +3,8 @@ import Button from '../button';
 import { Fragment, useState, useEffect } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-const { __ } = wp.i18n;
+import { __ } from '@wordpress/i18n';
+;
 
 import AmountControl from './amount-control';
 import PaymentMethodControl from './payment-method-control';