+++ src/DonorDashboards/resources/js/app/tabs/annual-receipts/content.js
@@ -1,6 +1,7 @@
 import { Fragment, useEffect } from 'react';
 
-const { __ } = wp.i18n;
+import { __ } from '@wordpress/i18n';
+;
 
 import Heading from '../../components/heading';
 import AnnualReceiptTable from '../../components/annual-receipt-table';